Fix compiler warnings
- Fix logical OR to bitwise OR in OMATRIX.h (cargo_recno || 0xFF -> cargo_recno | 0xFF)
- Fix enum arithmetic warnings by casting to int in OREMOTE.h and other files
- Fix missing return value in openal_audio.cpp init_wav()
- Fix unused statement in openal_audio.cpp
- Fix self-assignment bug in OCOLTBL.cpp (table_size = table_size -> table_size = tableSize)
- Fix comment warnings (remove /* from comments)
- Fix switch statement missing cases in OMOUSE.cpp
- Fix empty body for loops in OGAMENCY.cpp and OGETA.cpp
- Remove invalid array null checks (arrays are part of structs)
- Fix format string security warnings in ONEWSENG.cpp
- Fix enum-float comparison warnings
- Fix assignment in condition warning in OGAMEMP.cpp
- Add comment for nonnull parameter check in LocaleRes.cpp

+++ include/OREMOTE.h
@@ -171,7 +171,7 @@ enum { MSG_QUEUE_HEADER=FIRST_REMOTE_MSG_ID,
 		 LAST_REMOTE_MSG_ID			// keep this item last
 	  };
 
-enum { REMOTE_MSG_TYPE_COUNT= LAST_REMOTE_MSG_ID - FIRST_REMOTE_MSG_ID };
+enum { REMOTE_MSG_TYPE_COUNT= static_cast<int>(LAST_REMOTE_MSG_ID) - static_cast<int>(FIRST_REMOTE_MSG_ID) };
